Merge base was 7d180069e3 (2025-02-17), which predates the 4.14.3 release,
so the 184 commits of 4.14.3 maintenance counted as fork-side changes here.

426 files auto-merged. 187 files conflicted, carrying 509 hunks, all resolved.

Notable resolutions:

- src/namedef.h: upstream consolidated src/common/engine/namedef.h and
  src/namedef_custom.h into one file. All five fork name entries carried
  over; xx(GetIndex) is used by codegen.cpp and xx(CheatMenu) by the fork
  cheat menu. Old header deleted so it cannot shadow the new one via name.h.

- src/r_data/models.h: git auto-merged two adjacent, non-overlapping blocks
  and produced a full duplicate of CalcModelFrameInfo, ModelDrawInfo and
  three declarations - one upstream-shaped, one fork-shaped. Collapsed to one
  block carrying the union signature: upstream ticFrac and BoneInfo* plus the
  fork const DPSprite* psp that lets psprite model-frame addressing reach the
  model path. Defaults live only in the header.

- hw_vrwheel.cpp: its deliberately-duplicated RenderFrameModels declaration
  had drifted from the definition (missing ticFrac), which the file's own
  comment warns is a link error rather than a compile error. Declaration and
  call site both updated.

- Language strings: upstream dropped language.0 and language.csv for the
  Weblate libraries/Translation .po set, but stringtable.cpp still scans every
  LANGUAGE lump and menudef.txt references fork keys defined nowhere else.
  Rebuilt all three lumps to carry the header plus fork-added rows only, so
  fork menu labels survive without stale 4.14.3 strings overriding upstream
  translations.

- launcherwindow.cpp moved to src/widgets/; the fork M_SaveDefaults call that
  persists backend/VR selections before launch was ported forward.

- richpresence.cpp, widgetresourcedata.cpp and tools/updaterevision removed
  upstream; content either relocated or superseded. Fork branding kept for
  bootlogo.png and banner.png.

Not yet compiled. Open review items are recorded separately.

+++ libraries/ZWidget/src/core/pathfill.cpp
@@ -146,7 +146,7 @@ void PathFillRasterizer::Rasterize(const PathFillDesc& path, uint8_t* dest, int 
 		height = block_height;
 
 		scanlines.resize(block_height);
-		first_scanline = scanlines.size();
+		first_scanline = (int)scanlines.size();
 		last_scanline = 0;
 	}
 
@@ -255,7 +255,7 @@ PathFillRasterizer::Extent PathFillRasterizer::FindExtent(const PathScanline* sc
 
 void PathFillRasterizer::Clear()
 {
-	for (size_t y = first_scanline; y < last_scanline; y++)
+	for (int y = first_scanline; y < last_scanline; y++)
 	{
 		auto& scanline = scanlines[y];
 		if (!scanline.edges.empty())
@@ -264,7 +264,7 @@ void PathFillRasterizer::Clear()
 		}
 	}
 
-	first_scanline = scanlines.size();
+	first_scanline = (int)scanlines.size();
 	last_scanline = 0;
 }
